+++ server/apiRouter.js
@@ -19,6 +19,7 @@ const transitionPrivileged = require('./api/transition-privileged');
 const createUserWithIdp = require('./api/auth/createUserWithIdp');
 
 const { authenticateFacebook, authenticateFacebookCallback } = require('./api/auth/facebook');
+const { authenticateGoogle, authenticateGoogleCallback } = require('./api/auth/google');
 
 const router = express.Router();
 
@@ -68,5 +69,15 @@ router.get('/auth/facebook', authenticateFacebook);
 // with Facebook. In this route a Passport.js custom callback is used for calling
 // loginWithIdp endpoint in Flex API to authenticate user to Flex
 router.get('/auth/facebook/callback', authenticateFacebookCallback);
+
+// Google authentication endpoints
+
+// This endpoint is called when user wants to initiate authenticaiton with Google
+router.get('/auth/google', authenticateGoogle);
+
+// This is the route for callback URL the user is redirected after authenticating
+// with Google. In this route a Passport.js custom callback is used for calling
+// loginWithIdp endpoint in Flex API to authenticate user to Flex
+router.get('/auth/google/callback', authenticateGoogleCallback);
 
 module.exports = router;